KIEKER-1903 further cleanups.
