+++ kieker-analysis/src/kieker/analysis/architecture/dependency/AbstractDependencyGraphBuilder.java
@@ -18,7 +18,7 @@ package kieker.analysis.architecture.dependency;
 
 import java.time.temporal.ChronoUnit;
 
-import kieker.analysis.architecture.recovery.ModelRepository;
+import kieker.analysis.architecture.repository.ModelRepository;
 import kieker.analysis.graph.IEdge;
 import kieker.analysis.graph.IGraph;
 import kieker.analysis.graph.IVertex;
