+++ kieker-analysis/src/kieker/analysis/architecture/repository/ModelRepository.java
@@ -13,10 +13,41 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  ***************************************************************************/
+package kieker.analysis.architecture.repository;
+
+import java.util.HashMap;
+import java.util.Map;
+
+import org.eclipse.emf.ecore.EObject;
 
 /**
  * @author Reiner Jung
  * @since 1.15
  */
-package kieker.analysis.events.delayfilter.components;
+public class ModelRepository {
 
+	private final String name;
+
+	private final Map<Class<? extends EObject>, EObject> models = new HashMap<>();
+
+	public ModelRepository(final String name) {
+		this.name = name;
+	}
+
+	public Map<Class<? extends EObject>, EObject> getModels() {
+		return this.models;
+	}
+
+	@SuppressWarnings("unchecked")
+	public <T extends EObject> T getModel(final Class<T> clazz) {
+		return (T) this.models.get(clazz);
+	}
+
+	public void register(final Class<? extends EObject> key, final EObject value) {
+		this.models.put(key, value);
+	}
+
+	public String getName() {
+		return this.name;
+	}
+}
