+++ kieker-analysis/src/kieker/analysis/generic/RealtimeRecordDelayProducer.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  ***************************************************************************/
 
-package kieker.analysis.events.delayfilter.components;
+package kieker.analysis.generic;
 
 import java.util.concurrent.LinkedBlockingQueue;
 import java.util.concurrent.TimeUnit;
