+++ kieker-analysis/test/kieker/analysis/junit/plugin/filter/select/TestTimestampFilter.java
@@ -21,7 +21,7 @@ import java.nio.BufferOverflowException;
 import org.junit.Assert;
 import org.junit.Test;
 
-import kieker.analysis.architecture.generic.timestampfilter.TimestampFilter;
+import kieker.analysis.generic.timestampfilter.TimestampFilter;
 import kieker.common.record.flow.trace.AbstractTraceEvent;
 import kieker.common.record.io.IValueSerializer;
